refactor: PR 수정사항 최종반영
