+++ build.gradle
@@ -39,8 +39,6 @@ dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-oauth2-client'
     implementation 'org.springframework.boot:spring-boot-starter-security'
     implementation 'org.springframework.boot:spring-boot-starter-web'
-
-    // entity
     implementation 'org.springframework.boot:spring-boot-starter-validation'
 
     //db
