+++ src/main/java/site/sonisori/sonisori/common/Role.java
@@ -1,4 +1,4 @@
-package site.sonisori.sonisori.constants;
+package site.sonisori.sonisori.common;
 
 public enum Role {
 	ROLE_USER, ROLE_ADMIN
