+++ src/main/java/site/sonisori/sonisori/common/SocialType.java
@@ -1,4 +1,4 @@
-package site.sonisori.sonisori.constants;
+package site.sonisori.sonisori.common;
 
 public enum SocialType {
 	google, kakao
